[drake_bazel_...] Rewrite to actually be useful and correct
The objective of the sample projects within drake-external-examples is
to bootstrap our users. Users should be able to copy the sample
project into a new git tree and follow the samples instructions to
setup, build, and run.

We currently fail to meet this objective in a few important ways:

(1) The instructions were incomplete. They made assumptions about
additional steps that were not explained, or at least not clearly.

(2) The sample projects referred to files *outside of their tree*.
Using a top-level "scripts" folder that all of the samples refer to as
"../scripts" DEFEATS THE ENTIRE PURPOSE of the sample projects! When
the user copies the project to their own git tree, they do not have
access to those scripts anymore, and nothing works.

---

To fix those two problems, we make the following changes:

Split drake_bazel_installed into separate projects for "download" vs
"installed". Trying to show multiple workflows in a single directory
is too confusing.

Enhance drake_bazel_download to probe which OS it's running on, so
that it's suitable for any OS that Drake otherwise supports (i.e.,
don't hard-code it to focal anymore).

Fix drake_bazel_download to provide install_prereqs instead of relying
on the CI scripts. Not only are the CI scripts inaccessible once the
project is copied elsewhere, but they were badly implemented anyway
(for example, by hard-coding the list of Ubuntu packages, instead of
using the package list shipped alongside Drake).

Rename WORKSPACE to WORKSPACE.bazel for clarity.

Add github CI to the project starter code. Users shouldn't need to
figure out how to set up CI on their own; it should be part of the
build system starter pack.

This commit only fixes the `bazel_...` samples. The same problems
exist in all of the other examples as well, which will need to be
fixed in follow-up commits.

---

In support of those changes, we need to improve the maintainability of
these projects for the Drake development team:

Add file_sync_test to enforce that copies of files stay in sync.
TODO add this to CI actions

TODO Move linter dotfiles (cpplint, clang, etc.) down one level and
add to the file_sync_test.